feat(host-extension): extract WindowManager + RequestInterceptor (Phase 3b)
Largest of the extension extractions: ~740-line WindowManager class
plus the webRequest listener setup it depends on.

WindowManager
- Moved as-is to packages/host-extension/src/background/WindowManager.ts.
  Public surface unchanged: registerWindow, closeWindow, getWindow,
  getWindowByTabId, addRequest, addHeader, showOverlay, showPluginUI,
  reRenderPluginUI, cleanupInvalidWindows, plus the message-batching
  internals and the 10-window MAX limit.
- Internal imports rewritten to relative ../types and ../util paths now
  that the file lives inside the package.
- The associated 42-test suite (tests/background/WindowManager.test.ts)
  moves with it and continues to cover lifecycle + request batching
  scenarios.

RequestInterceptor (new export)
- Pulled the inline webRequest.onBeforeRequest +
  webRequest.onBeforeSendHeaders listener blocks out of
  packages/extension/src/entries/Background/index.ts into a small
  `installRequestInterceptor({ windowManager })` helper in the package.
  Both listeners exit early when getWindowByTabId returns nothing, so
  the scoping behavior (broad <all_urls> permission limited to plugin-
  opened windows) is preserved.
- Background/index.ts is now ~50 lines lighter — it just calls
  installRequestInterceptor() instead of inlining the listener bodies,
  and drops the now-unused InterceptedRequest / InterceptedRequestHeader
  type imports.

Exports
- @tlsn/host-extension/background now exports `WindowManager`,
  `installRequestInterceptor`, and `ExtensionRequestInterceptorOptions`.

Verification
- host-extension: 133/133 tests pass (gained 42 from WindowManager.test)
- extension: 51/51 tests pass (lost WindowManager.test but the rest
  continue green)
- Both packages still build clean.

Next slices: SessionManager + ProveManager (the offscreen prover
internals); ConfirmationManager + ApprovalUi; the content-script
PluginRenderer.
