+++ packages/host-extension/src/background/index.ts
@@ -1,0 +1,5 @@
+export { WindowManager } from './WindowManager.js';
+export {
+  installRequestInterceptor,
+  type ExtensionRequestInterceptorOptions,
+} from './RequestInterceptor.js';
